# docs/swagger.json
{
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "API RESTful de exemplo com documentação Swagger automática",
    "title": "API de Exemplo em Go",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "Suporte API",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
    "/health": {
      "get": {
        "description": "Retorna o status de saúde do servidor",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Health"
        ],
        "summary": "Verificar saúde da API",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "description": "Retorna a lista de todos os usuários cadastrados",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Listar todos os usuários",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.User"
              }
            }
          },
          "500": {
            "description": "Erro ao codificar resposta",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "post": {
        "description": "Cria um novo usuário no sistema",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Criar novo usuário",
        "parameters": [
          {
            "description": "Dados do Usuário",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          },
          "400": {
            "description": "Dados inválidos",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Erro ao codificar resposta",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Retorna um usuário específico pelo ID",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Buscar usuário por ID",
        "parameters": [
          {
            "type": "integer",
            "description": "ID do Usuário",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          },
          "400": {
            "description": "ID inválido",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Erro ao codificar resposta",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "models.User": {
      "type": "object",
      "properties": {
        "age": {
          "type": "integer",
          "example": 30
        },
        "created_at": {
          "type": "string",
          "example": "2026-01-09T10:00:00Z"
        },
        "email": {
          "type": "string",
          "example": "joao@example.com"
        },
        "id": {
          "type": "integer",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "João Silva"
        },
        "updated_at": {
          "type": "string",
          "example": "2026-01-09T10:00:00Z"
        }
      }
    }
  }
}